+++ BackEnd/My_Eshop/src/main/java/com/example/my_eshop/entity/DTO/DtoOrder.java
@@ -1,4 +1,4 @@
-package com.example.my_eshop.order.DTO;
+package com.example.my_eshop.entity.DTO;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
