+++ BackEnd/My_Eshop/src/main/java/com/example/my_eshop/entity/Orderline.java
@@ -1,7 +1,5 @@
-package com.example.my_eshop.orderline;
+package com.example.my_eshop.entity;
 
-import com.example.my_eshop.order.Order;
-import com.example.my_eshop.product.Product;
 import jakarta.persistence.*;
 import lombok.Data;
