+++ BackEnd/My_Eshop/src/main/java/com/example/my_eshop/repository/OrderRepository.java
@@ -1,5 +1,6 @@
-package com.example.my_eshop.order;
+package com.example.my_eshop.repository;
 
+import com.example.my_eshop.entity.Order;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface OrderRepository extends JpaRepository<Order,Integer> {
