+++ BackEnd/My_Eshop/src/main/java/com/example/my_eshop/repository/OrderlineRepository.java
@@ -1,5 +1,6 @@
-package com.example.my_eshop.orderline;
+package com.example.my_eshop.repository;
 
+import com.example.my_eshop.entity.Orderline;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface OrderlineRepository extends JpaRepository<Orderline,Integer> {
