+++ BackEnd/My_Eshop/src/main/java/com/example/my_eshop/repository/ProductRepository.java
@@ -1,5 +1,6 @@
-package com.example.my_eshop.product;
+package com.example.my_eshop.repository;
 
+import com.example.my_eshop.entity.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ProductRepository extends JpaRepository<Product,Integer> {
