+++ BackEnd/My_Eshop/src/main/java/com/example/my_eshop/repository/UserRepository.java
@@ -1,5 +1,6 @@
-package com.example.my_eshop.user;
+package com.example.my_eshop.repository;
 
+import com.example.my_eshop.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface UserRepository extends JpaRepository<User,Integer> {
