+++ BackEnd/My_Eshop/src/main/java/com/example/my_eshop/service/interfaces/IUserService.java
@@ -1,4 +1,6 @@
-package com.example.my_eshop.user;
+package com.example.my_eshop.service.interfaces;
+
+import com.example.my_eshop.entity.User;
 
 import java.util.List;
 import java.util.Optional;
