+++ BackEnd/My_Eshop/src/main/resources/application.properties
@@ -4,9 +4,9 @@
 
 
 spring.jpa.hibernate.ddl-auto=create-drop
-spring.datasource.url=jdbc:mysql://localhost:3306/new_db_my_eshop
+spring.datasource.url=jdbc:mysql://localhost:3306/testdb
 spring.datasource.username=root
-spring.datasource.password=HugoHerich9394+
+spring.datasource.password=rootroot
 spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
 spring.jpa.hibernate.naming.physical-strategy=org.hibernate.boot.model.naming.PhysicalNamingStrategyStandardImpl
